refactor

+++ DTO/Training/TrainingDto.cs
@@ -20,7 +20,7 @@ public class TrainingDto
     public record StrengthExerciseDto(int ExerciseId, IEnumerable<StrengthParamsDto> Params);
     public ICollection<StrengthExerciseDto>? StrengthExercises { get; set; }
     
-    public record CardioParamsDto(int Interval, int Speed, TimeSpan Time);
+    public record CardioParamsDto(int Interval, double Speed, TimeSpan Time);
     public record CardioExerciseDto(int ExerciseId, IEnumerable<CardioParamsDto> Params);
     public ICollection<CardioExerciseDto>? CardioExercises { get; set; }
 }
